# feat: faster way of figuring out how many new scraper rows were inserted